feat(消息处理): 增强消息类型支持结构化消息和批量操作
- 在Messages枚举中新增Structured变体以支持结构化消息
- 为Messages类型添加验证和转换方法(to_content/to_message_list)
- 实现批量内存操作处理逻辑
- 添加enhanced-messages-demo示例展示新功能
- 清理未使用的导入和优化代码结构

+++ crates/agent-mem-llm/src/providers/together.rs
@@ -284,7 +284,7 @@ impl LLMProvider for TogetherProvider {
         ModelInfo {
             model: self.config.model.clone(),
             provider: "together".to_string(),
-            max_tokens: self.get_model_max_tokens(),
+            max_tokens: self.get_model_max_tokens() as u32,
             supports_streaming: false, // 流式需要额外实现
             supports_functions: false, // Together 目前不支持函数调用
         }
